--- src/js/main.js
@@ -1,6 +1,0 @@
-import productList from "./productList.mjs";
-import loadHeaderFooter from "/utils.mjs";
-
-loadHeaderFooter();
-
-productList(".product-list", "tents");
