+++ src/public/checkout/index.html
