+++ src/public/js/product.js
@@ -1,10 +1,6 @@
-import { getParam, setLocalStorage } from "./utils.mjs";
+import { setLocalStorage } from "./utils.mjs";
 import { getLocalStorage } from "./utils.mjs";
 import { findProductById } from "./productData.mjs";
-import productDetails from "./productDetails.mjs";
-
-const productId = getParam("product")
-productDetails(productId);
 
 function addProductToCart(product) {
   let cartItems = getLocalStorage("so-cart", product);
