--- src/public/partials/header.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    <header class="divider" id="main-header">
-  
-    </header>
-    <div class="logo">
-        <img src="/images/noun_Tent_2517.svg" alt="tent image for logo">
-        <a href="index.html"> Sleep<span class="highlight">Outside</span></a>
-      </div>
-      <div class="cart">
-        <a href="cart.html">
-          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 100 100">
-            ...all the SVG code ...
-          </svg>
-        </a>
-      </div>
-</body>
-</html>
