+++ team.txt
@@ -3,4 +3,3 @@ https://trello.com/b/cY5DzmJ3/wdd-team-project
 Garrett
 Benjamin
 Hayden
-Andy
