+++ vite.config.js
@@ -10,9 +10,20 @@ export default defineConfig({
       input: {
         main: resolve(__dirname, "src/index.html"),
         cart: resolve(__dirname, "src/cart/index.html"),
-        product: resolve(__dirname, "src/product_pages/index.html"
+        checkout: resolve(__dirname, "src/checkout/index.html"),
+        product1: resolve(
+          __dirname,
+          "src/product_pages/cedar-ridge-rimrock-2.html"
         ),
-        
+        product2: resolve(__dirname, "src/product_pages/marmot-ajax-3.html"),
+        product3: resolve(
+          __dirname,
+          "src/product_pages/northface-alpine-3.html"
+        ),
+        product4: resolve(
+          __dirname,
+          "src/product_pages/northface-talus-4.html"
+        ),
       },
     },
   },
